HMAN-56 Fixing/moving unit tests wip
