+++ src/test/java/uk/gov/hmcts/reform/hmc/validator/HearingIdValidatorTest.java
@@ -9,12 +9,14 @@ import uk.gov.hmcts.reform.hmc.data.HearingEntity;
 import uk.gov.hmcts.reform.hmc.data.LinkedGroupDetails;
 import uk.gov.hmcts.reform.hmc.domain.model.enums.PutHearingStatus;
 import uk.gov.hmcts.reform.hmc.exceptions.BadRequestException;
+import uk.gov.hmcts.reform.hmc.exceptions.HearingNotFoundException;
 import uk.gov.hmcts.reform.hmc.repository.HearingRepository;
 
 import java.util.Optional;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertThrows;
+import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.when;
 import static uk.gov.hmcts.reform.hmc.exceptions.ValidationError.INVALID_HEARING_ID_DETAILS;
 
@@ -57,6 +59,50 @@ class HearingIdValidatorTest {
     void shouldSucceedAsValidHearingIdFormat() {
         when(hearingRepository.existsById(VALID_HEARING_ID)).thenReturn(true);
         hearingIdValidator.validateHearingId(VALID_HEARING_ID, null);
+    }
+
+    @Test
+    void updateHearingRequestShouldThrowErrorWhenHearingIdNotPresentInDB() {
+        final String expectedErrorMsg = "Whatever error message";
+        when(hearingRepository.existsById(any())).thenReturn(false);
+        Exception exception = assertThrows(HearingNotFoundException.class, () -> hearingIdValidator
+                .validateHearingId(2000000000L, expectedErrorMsg));
+        assertEquals(expectedErrorMsg, exception.getMessage());
+    }
+
+    @Test
+    void updateHearingRequestShouldThrowErrorWhenHearingIdDoesNotStartWith2() {
+        Exception exception = assertThrows(BadRequestException.class, () -> hearingIdValidator
+                .validateHearingId(1000000100L, INVALID_HEARING_ID_DETAILS));
+        assertEquals(INVALID_HEARING_ID_DETAILS, exception.getMessage());
+    }
+
+    @Test
+    void updateHearingRequestShouldThrowErrorWhenHearingIdExceedsMaxLength() {
+        Exception exception = assertThrows(BadRequestException.class, () -> hearingIdValidator
+                .validateHearingId(20000000001111L, INVALID_HEARING_ID_DETAILS));
+        assertEquals(INVALID_HEARING_ID_DETAILS, exception.getMessage());
+    }
+
+    @Test
+    void updateHearingRequestShouldThrowErrorWhenHearingIdIsNull() {
+        Exception exception = assertThrows(BadRequestException.class, () -> hearingIdValidator
+                .validateHearingId(null, INVALID_HEARING_ID_DETAILS));
+        assertEquals(INVALID_HEARING_ID_DETAILS, exception.getMessage());
+    }
+
+    @Test
+    void testExpectedException_DeleteHearing_HearingId_Exceeds_MaxLength() {
+        Exception exception = assertThrows(BadRequestException.class, () -> hearingIdValidator
+                .validateHearingId(20000000001111L, "Invalid hearing Id"));
+        assertEquals("Invalid hearing Id", exception.getMessage());
+    }
+
+    @Test
+    void testExpectedException_DeleteHearing_HearingId_First_Char_Is_Not_2() {
+        Exception exception = assertThrows(BadRequestException.class, () -> hearingIdValidator
+                .validateHearingId(1000000100L, "Invalid hearing Id"));
+        assertEquals("Invalid hearing Id", exception.getMessage());
     }
 
     private HearingEntity generateHearing(Long id, String status, LinkedGroupDetails groupDetails, Long linkedOrder) {
